Revert "Prevent integer underflow in `CollisionResolver::update`"
This reverts commit 4c7aa4fde7c6faa1fd46ff02a9bf050e7cf39ac2.

+++ src/ECS/CollisionResolverComponent.cpp
@@ -28,8 +28,7 @@ void CollisionResolver::update() {
           if ((transform->position.x + transform->width / 2) -
                   (coll->collider.x + coll->collider.w / 2) <
               0) {
-            transform->position.x =
-                static_cast<float>(coll->collider.x) - transform->width;
+            transform->position.x = coll->collider.x - transform->width;
           } else {
             transform->position.x = coll->collider.x + coll->collider.w;
           }
@@ -38,8 +37,7 @@ void CollisionResolver::update() {
         if ((transform->position.x + transform->width / 2) -
                 (coll->collider.x + coll->collider.w / 2) <
             0) {
-          transform->position.x =
-              static_cast<float>(coll->collider.x) - transform->width;
+          transform->position.x = coll->collider.x - transform->width;
         } else {
           transform->position.x = coll->collider.x + coll->collider.w;
         }
@@ -47,8 +45,7 @@ void CollisionResolver::update() {
         if ((transform->position.y + transform->height / 2) -
                 (coll->collider.y + coll->collider.h / 2) <
             0) {
-          transform->position.y =
-              static_cast<float>(coll->collider.y) - transform->height;
+          transform->position.y = coll->collider.y - transform->height;
         } else {
           transform->position.y = coll->collider.y + coll->collider.h;
         }
